requirements.txt
    gradio
    pymilvus
    sentence-transformers
    openai
    langchain
    python-dotenv
    langchain-community
    langchain_openai